backend/requirements.txt
# Django Backend Requirements
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
celery>=5.3.0
redis>=4.5.0
django-redis>=5.2.0
Pillow>=10.0.0
gunicorn>=21.0.0

# Development dependencies
django-debug-toolbar>=4.1.0
pytest>=7.4.0
pytest-django>=4.5.0
black>=23.0.0
flake8>=6.0.0
